requirements.txt
Django==1.9
psycopg2==2.7.3.1
django-rest-framework==0.1.0
